+++ packages/reactive-component-loader/src/lib/lazy/lazy.component.ts
@@ -1,6 +1,5 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { Observable } from 'rxjs';
-import { tap } from 'rxjs/operators';
 import { ComponentLocation, ComponentRecipe, ReactiveComponentLoader } from '../reactive-component-loader.service';
 import { Inputs } from './inputs';
 import { Outputs } from './outputs';