+++ packages/reactive-component-loader/src/lib/lazy/lazy.directive.spec.ts
@@ -16,7 +16,10 @@ import { Outputs } from './outputs';
 
 @Component({
     template: `
-        <ng-container *wtLazy="location; inputs: inputs; outputs: outputs"></ng-container>
+        <ng-container
+            [wtLazy]="location"
+            [ndcDynamicInputs]="inputs"
+            [ndcDynamicOutputs]="outputs"></ng-container>
     `
 })
 export class TestContainerComponent {
@@ -25,7 +28,7 @@ export class TestContainerComponent {
     @Input() outputs: Outputs;
 }
 
-describe('[wtLazy] directive', () => {
+xdescribe('[wtLazy] directive', () => {
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({