+++ packages/reactive-component-loader/src/lib/reactive-component-loader.module.ts
@@ -7,12 +7,12 @@
 
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule, NgModuleFactoryLoader, SystemJsNgModuleLoader } from '@angular/core';
-import { provideRoutes, Route } from '@angular/router';
+import { provideRoutes } from '@angular/router';
 import { DynamicModule } from 'ng-dynamic-component';
 import { REACTIVE_COMPONENT_LOADER_MODULE_REGISTRY } from './_internals';
 import { LazyComponent } from './lazy/lazy.component';
-import { ModuleInfo } from './module-info';
 import { LazyDirective } from './lazy/lazy.directive';
+import { ModuleInfo } from './module-info';
 
 export function noMatch() {
     return null;
@@ -48,6 +48,7 @@ export function noMatch() {
         LazyDirective
     ],
     exports: [
+        DynamicModule,
         LazyComponent,
         LazyDirective
     ],